+++ CHANGELOG.md
@@ -35,6 +35,7 @@
   - `Rsort()` now supports tracing
   - `Sort()` now supports tracing
   - `StripExtension()` now supports tracing
+  - `SwapExtensions()` now supports tracing
   - `TestEmpty()` now supports tracing
   - `TestFilepathExists()` now supports tracing
   - `TestNotEmpty()` now supports tracing